% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{comm_uni}
\alias{comm_uni}
\title{Common and Unique regions}
\usage{
comm_uni(data_1, data_2, pop, class)
}
\arguments{
\item{data_1}{A file obtained from the functions ```get_ROHi()``` or ```get_RHZ()``` for the first population.}

\item{data_2}{A file obtained from the functions ```get_ROHi()``` or ```get_RHZ()``` for the second population.}

\item{pop}{target population you want to get the outcome, in quotation marks ("").}

\item{class}{Type of outcomes we want. Two choices: "Unique" or "Common". n quotation marks ("").}
}
\value{
Pairwise common and Unique regions
}
\description{
This function finds the unique and common ROH islands or regions of heterozygosity between two
populations.
}
\examples{
# Obtaining common and unique ROHi for Europe and Africa:
eu<-subset(ROHi_HGDP,ROHi_HGDP$Population=="Europe")
af<-subset(ROHi_HGDP,ROHi_HGDP$Population=="Africa")
eur_uni<-comm_uni(eu,af,"Europe","Unique")
afr_uni<-comm_uni(eu,af,"Africa","Unique")
eur_uni
afr_uni
}
