% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID_genes}
\alias{COVID_genes}
\title{Protein coding genes in COVID ROHi}
\format{
A data frame with 153 rows and 7 variables.
\describe{
  \item{ensembl_gene_id}{Gene ID Ensembl format}
  \item{chromosome_name}{Chromosome number}
  \item{start_position}{start of the gene}
  \item{end_position}{end of the gene}
  \item{gene_biotype}{gene type_ protein coding gene}
  \item{external_gene_name}{gene name}
  \item{description}{description of the gene}
}
}
\usage{
COVID_genes
}
\description{
This dataset contains the protein coding genes found in the ROH islands of the COVID 19 dataset
}
\keyword{datasets}
