% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID_ID}
\alias{COVID_ID}
\title{Inbreeding Depression analysis with COVID-19 data.}
\format{
A data frame with 7387 rows and 14 variables.
\describe{
  \item{a1hosp}{Hospitalization 0 or 1}
  \item{Froh}{Inbreeding genomic coefficient}
  \item{Fhat1}{Fgrm}
  \item{age}{age of the individual}
  \item{pc1}{Pricipal component 1}
  \item{pc2}{Pricipal component 2}
  \item{pc3}{Pricipal component 3}
  \item{pc4}{Pricipal component 4}
  \item{pc5}{Pricipal component 5}
  \item{pc6}{Pricipal component 6}
  \item{pc7}{Pricipal component 7}
  \item{pc8}{Pricipal component 8}
  \item{pc9}{Pricipal component 9}
  \item{pc10}{Pricipal component 10}
}
}
\usage{
COVID_ID
}
\description{
This dataset contains the information needed to perform an inbreeding depression analysis with hospitalization during COVID-19.
}
\keyword{datasets}
