% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID_ROHi}
\alias{COVID_ROHi}
\title{ROH islands in COVID ROHi}
\format{
A data frame with 153 rows and 7 variables.
\describe{
  \item{Chr}{Chromosome number}
  \item{Start}{start of the ROH island}
  \item{End}{end of the ROH island}
  \item{N_Individuals}{number of individuals}
  \item{Lenght}{length of the ROHisland}
}
}
\usage{
COVID_ROHi
}
\description{
This dataset contains the protein coding genes found in the ROH islands of the COVID 19 dataset
}
\keyword{datasets}
