% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{genomic_repre}
\alias{genomic_repre}
\title{Genomic representation of regions}
\usage{
genomic_repre(data_1, data_2, pop)
}
\arguments{
\item{data_1}{A file obtained from the functions ```get_ROHi()```.}

\item{data_2}{A file obtained from the functions ```get_RHZ()```.}

\item{pop}{population to be represented.}
}
\value{
Figure of the ROHi and RHZ in a popualtion
}
\description{
This script creates a figure of the Number of ROH>=1.5Mb vs. Sum of ROH>=1.5Mb.
It is possible to add the simulated number and sum of ROH for different consanguineous mating.
The dashed diagonal line represents the regression line of N vs S of ROH for two admixed populations from the 1K genomes: ACB and ASW
}
