% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{get_ROH_deserts}
\alias{get_ROH_deserts}
\title{ROH deserts (regions outside ROH islands) by fixed prevalence threshold}
\usage{
get_ROH_deserts(
  POP,
  ChroNumber,
  population,
  thr = 0.01,
  SizeWindow = 10000,
  lenChro = NULL,
  min_windows = 1
)
}
\arguments{
\item{POP}{A .hom file/data.frame from PLINK (single population)}

\item{ChroNumber}{Chromosome number}

\item{population}{Name of the population}

\item{thr}{Fixed threshold on prevalence (fraction). Example: 0.01 = 1% of individuals}

\item{SizeWindow}{Window size (bp), default 10kb}

\item{lenChro}{Optional chromosome length (bp). If NULL uses your internal table}

\item{min_windows}{Minimum consecutive windows to report a region}
}
\value{
A table of ROH desert regions
}
\description{
ROH deserts (regions outside ROH islands) by fixed prevalence threshold
}
