% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HGDP_cont}
\alias{HGDP_cont}
\title{Continents in Human Genome Diversity Panel.}
\format{
A data frame with 57 rows and 2 variables:
\describe{
  \item{pop}{population}
  \item{cont}{continent}
}
}
\source{
\url{https://cephb.fr/en/hgdp_panel.php}
}
\usage{
HGDP_cont
}
\description{
This dataset contains the classification in continents for all the populations in the HGD dataset:
}
\keyword{datasets}
