% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HGDP_hom}
\alias{HGDP_hom}
\title{ROH in the Human Genome Diversity Panel.}
\format{
## HGDP_hom
A data frame with 612139 rows and 13 variables:
\describe{
  \item{IID}{individual id}
  \item{FID}{family id}
  \item{PHE}{Phenotype not present}
  \item{CHR}{Chromosome number}
  \item{SNP1}{rs of the SNP in starting possition}
  \item{SNP2}{rs of the SNP in ending possition}
  \item{POS1}{starting of the ROH}
  \item{POS2}{ending of the ROH}
  \item{KB}{ROH length in KB}
  \item{NSNP}{Number of SNP present in the ROH}
  \item{DENSITY}{Densito of SNP}
  \item{PHOM}{}
  \item{PHET}{}
}
}
\source{
\url{https://cephb.fr/en/hgdp_panel.php}
}
\usage{
HGDP_hom
}
\description{
This dataset contains the PLINKs .hom file obtained using Human Genome Diversity Panel.
PLINKs parameters were set to:
homozyg-window-snp 30.
homozyg-window-het 1.
homozyg-snp 30.
homozyg-kb 300.
homozyg-density 30.
}
\keyword{datasets}
