% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HGDP_pops}
\alias{HGDP_pops}
\title{Populations Human Genome Diversity Panel.}
\format{
A data frame with 57 rows and 2 variables:
\describe{
  \item{IID}{individual id}
  \item{pop}{population}
}
}
\source{
\url{https://cephb.fr/en/hgdp_panel.php}
}
\usage{
HGDP_pops
}
\description{
This dataset contains the classification in populations for all the individuals in the 1K Genomes dataset:
}
\keyword{datasets}
