% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{in_dep_log}
\alias{in_dep_log}
\title{INBREEDING DEPRESSION: BINOMIAL DEPENDENT VARIABLE.}
\usage{
in_dep_log(form, df)
}
\arguments{
\item{form}{model formula. For example: `dep_var~indep_var1+indep_var2+indep_var3`.}

\item{df}{D.B_Final database obtained from the data incorporation step.}
}
\value{
A table with the Odds ratio, its 95% CI and the associated p-value.
}
\description{
The purpose of this function is to calculate the inbreeding depression of a dichotomous dependent trait with multiple independent variables of any nature.
The inbreeding depression is given as an odds ratio (OR).
}
\examples{
# Basic Usage of Multivariable logistic regression.
in_dep_log(a1hosp~Froh+Fhat1+age+pc1+pc2+pc3+pc4+pc5+pc6+pc7+pc8+pc9+pc10,COVID_ID)

}
