% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KGenomes_cont}
\alias{KGenomes_cont}
\title{Continents in 1K Genomes.}
\format{
A data frame with 57 rows and 2 variables:
\describe{
  \item{pop}{population}
  \item{cont}{continent}
}
}
\source{
\url{http://https://www.internationalgenome.org}
}
\usage{
KGenomes_cont
}
\description{
This dataset contains the classification in continents for all the populations in the 1K Genomes dataset:
}
\keyword{datasets}
