% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KGenomes_het}
\alias{KGenomes_het}
\title{FIS in 1K Genomes.}
\format{
A data frame with 936 rows and 2 variables:
\describe{
  \item{IID}{individual id}
  \item{Fis}{FIS estimate obtained using PLINK's --het flag}
}
}
\source{
\url{http://https://www.internationalgenome.org}
}
\usage{
KGenomes_het
}
\description{
This dataset contains the average FIS stimates for each individual in the 1K Genomes dataset:
}
\keyword{datasets}
