% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KGenomes_pops}
\alias{KGenomes_pops}
\title{Populations 1K Genomes.}
\format{
A data frame with 57 rows and 2 variables:
\describe{
  \item{IID}{individual id}
  \item{pop}{population}
}
}
\source{
\url{http://https://www.internationalgenome.org}
}
\usage{
KGenomes_pops
}
\description{
This dataset contains the classification in populations for all the individuals in the 1K Genomes dataset:
}
\keyword{datasets}
