% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{n_vs_sum}
\alias{n_vs_sum}
\title{Number vs Sum of ROH}
\usage{
n_vs_sum(data_1, data_2, simul = TRUE)
}
\arguments{
\item{data_1}{The outcome of the function roh_sum_id.}

\item{data_2}{A file with two columns: IID, pop. pop must contain each individula's population}

\item{simul}{If true simulations of number and sum of ROH for differnet consanguinity matings is added.}
}
\value{
A figure of the Number vs the Sum of ROH for ROH larger than 1.5Mb.
}
\description{
This script creates a figure of the Number of ROH>=1.5Mb vs. Sum of ROH>=1.5Mb.
It is possible to add the simulated number and sum of ROH for different consanguineous mating.
The dashed diagonal line represents the regression line of N vs S of ROH for two admixed populations from the 1K genomes: ACB and ASW
}
