% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RHZ_HGDP}
\alias{RHZ_HGDP}
\title{Runs of Heterozygosity RHZ in Human Genome Diversity Panel.}
\format{
A data frame with 3104 rows and 7 variables.
\describe{
  \item{Chr}{Chromosome}
  \item{Start}{Starting possition of the RHZ}
  \item{End}{Ending possition of the RHZ}
  \item{Length}{Length in Mb of the RHZ}
  \item{N_Individuals}{Number of individuals that do not have a ROH in that genomic location, so its in RHZ}
  \item{P_Individuals}{Proportion, in \%, of the individuals in the sample that do not have a ROH in that genomic location}
  \item{Population}{Name of the population}
}
}
\usage{
RHZ_HGDP
}
\description{
This dataset contains the Runs of Heterozygosity of different groups of the Human Genome Diversity Panel.
}
\keyword{datasets}
