% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rhz_summ_pop}
\alias{rhz_summ_pop}
\title{Summarize RHZ by Population}
\usage{
rhz_summ_pop(mypath)
}
\arguments{
\item{mypath}{Path to the folder with the outcomes of the ```get_RHZ()``` function.}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This script summarize all the outcome of RHZ by population.
}
