% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rhz_summ_pop_file}
\alias{rhz_summ_pop_file}
\title{Summarize RHZ by Population and by file}
\usage{
rhz_summ_pop_file(dat)
}
\arguments{
\item{mypath}{Path to the folder with the outcomes of the ```get_RHZ()``` function.}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This script summarize all the outcome of RHZ by population.
}
\examples{
# Summarize ROH islands for a group of individuals:
rhz_summ_pop_file(RHZ_HGDP)

}
