% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{roh_sum_pop}
\alias{roh_sum_pop}
\title{Summarize by Population}
\usage{
roh_sum_pop(data_1, data_2)
}
\arguments{
\item{data_1}{The outcome of the function roh_sum_id.}

\item{data_2}{A file with two columns: IID, pop. pop must contain each individula's population}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This script summarize all the outcome of roh_sum_id by population.
}
