% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{roh_summ_cont}
\alias{roh_summ_cont}
\title{Summarize by Continent}
\usage{
roh_summ_cont(data_1, data_2)
}
\arguments{
\item{data_1}{The outcome of the function ```roh_summ_id()```.}

\item{data_2}{A file with two columns: "IID", "cont". cont must contain each individual's continent}
}
\value{
A data frame with different variables summarize for each continent.
}
\description{
This script summarize all the outcome of ```roh_summ_id()``` by population.
}
\examples{
# Obtaining basic ROH variables:
rohsum<-roh_summ_id(KGenomes_hom)
pops<-merge(KGenomes_pops,KGenomes_cont,by="pop")
roh_summ_cont(rohsum,pops)

}
