% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{roh_summ_factor}
\alias{roh_summ_factor}
\title{Summarize by factor}
\usage{
roh_summ_factor(data_1, data_2, group_vars)
}
\arguments{
\item{data_1}{The outcome of the function ```roh_summ_id()```.}

\item{data_2}{A file with each individual's factor}
}
\value{
A data frame with different variables summarize for each continent.
}
\description{
This script summarize all the outcome of ```roh_summ_id()``` by any factor.
}
