% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{roh_summ_pop_short}
\alias{roh_summ_pop_short}
\title{Summarize by Population, Short version}
\usage{
roh_summ_pop_short(data_1, data_2)
}
\arguments{
\item{data_1}{The outcome of the function ```roh_summ_id()```.}

\item{data_2}{A file with two columns: "IID", "pop". pop must contain each individual's population}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This function summarize all the outcome of ```roh_summ_id()``` by population.
}
\examples{
# Summarizing ROH variables:
rohsum<-roh_summ_id(HGDP_hom)
roh_summ_pop_short(rohsum,HGDP_pops)

}
