% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rohi_genome}
\alias{rohi_genome}
\title{ROHi Representation}
\usage{
rohi_genome(data_1, pop)
}
\arguments{
\item{data_1}{A file obtained from the functions ```get_ROHi()``` or ```get_RHZ()``` for a population}

\item{pop}{target population you want to get the outcome, in quotation marks ("").}
}
\value{
Figure of the ROHi or RHZ in a popualtion
}
\description{
This script creates a genomic representation of the ROH islands or Regions of Heterozygosity (RHZ).
}
