% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROHi_HGDP}
\alias{ROHi_HGDP}
\title{ROH islands in Human Genome Diversity Panel.}
\format{
A data frame with 3104 rows and 7 variables.
\describe{
  \item{Chr}{Chromosome}
  \item{Start}{Starting possition of the ROH island}
  \item{End}{Ending possition of the ROH island}
  \item{Length}{Length in Mb of the ROH island}
  \item{N_Individuals}{Number of individuals that have a ROH in that genomic location}
  \item{P_Individuals}{Proportion, in \%, of the individuals in the sample that have a ROH in that genomic location}
  \item{Population}{Name of the population}
}
}
\usage{
ROHi_HGDP
}
\description{
This dataset contains the ROH islands of different groups of the Human Genome Diversity Panel.
}
\keyword{datasets}
