% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rohi_rohz_genome}
\alias{rohi_rohz_genome}
\title{ROHi + ROHZ Representation}
\usage{
rohi_rohz_genome(data_1, data_2, pop)
}
\arguments{
\item{data_1}{A file obtained from the functions ```get_ROHi()```.}

\item{data_2}{A file obtained from the functions ```get_RHZ()```.}

\item{pop}{population to be represented.}
}
\value{
Figure of the ROHi and RHZ in a popualtion
}
\description{
This script creates a genomic representation of ROH islands and regions of heterozygosity (RHZ).
}
