% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rohi_summ_pop}
\alias{rohi_summ_pop}
\title{Summarize ROHi by Population}
\usage{
rohi_summ_pop(mypath)
}
\arguments{
\item{mypath}{Path to the folder with the outcomes of the get_RHOi() function.}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This script summarize all the outcome of get_ROHi by population.
}
