% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rohi_summ_pop_file}
\alias{rohi_summ_pop_file}
\title{Summarize ROHi by Population and by file}
\usage{
rohi_summ_pop_file(dat)
}
\arguments{
\item{dat}{DF with the outcomes of the ```get_RHOi()``` function for different populations.}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This script summarize all the outcome of get_ROHi by population.
}
\examples{
# Summarize ROH islands for a group of individuals:
rohi_summ_pop_file(ROHi_HGDP)

}
