% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rohi_summ_pop_path}
\alias{rohi_summ_pop_path}
\title{Summarize ROHi by Population and by folder}
\usage{
rohi_summ_pop_path(mypath)
}
\arguments{
\item{mypath}{Path to the folder with the outcomes of the ```get_RHOi()``` function.}
}
\value{
A data frame with different variables summarize for each population.
}
\description{
This script summarize all the outcome of get_ROHi by population.
}
