% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{get_Prot}
\alias{get_Prot}
\title{Searching for Protein Coding Genes}
\usage{
get_Prot(DATA)
}
\arguments{
\item{DATA}{A file obtained from the functions get_ROHi or get_RHZ.}
}
\value{
A table with the Protein coding genes
}
\description{
This script harvest protein coding genes from the ROHi or RHZ regions. It uses biomart package
}
