% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{get_RHOi}
\alias{get_RHOi}
\title{Islands of ROH}
\usage{
get_RHOi(POP, ChroNumber)
}
\arguments{
\item{POP}{A .hom file from PLINK with all the individuals belonging to the same group or population.}

\item{ChroNumber}{Chromosome number}
}
\value{
A table with the ROH islands
}
\description{
This script searches for ROH islands in a population
}
