% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{get_RHOi}
\alias{get_RHOi}
\title{Islands of ROH}
\usage{
get_RHOi(POP, ChroNumber, population)
}
\arguments{
\item{POP}{A .hom file from PLINK with all the individuals belonging to the same group or population.}

\item{ChroNumber}{Chromosome number.}

\item{population}{Name of the population in quotation marks ("").}
}
\value{
A table with the ROH islands.
}
\description{
This script searches for ROH islands in a population. The script may take some time to finish the 22 Chr.
The resolution of the analysis is 10,000 bp: 10kb.
}
\examples{
# Obtaining ROH islands for a group of individuals:
get_RHOi(HGDP_cl_Africa,1,"Africa")

}
