% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{get_RHZ}
\alias{get_RHZ}
\title{RUNS OF HETEROZYGOSITY II}
\usage{
get_RHZ(DF, ChroNumber, PR, population)
}
\arguments{
\item{DF}{The data frame obtained from the function ```rhc_data_org()```.}

\item{ChroNumber}{Chromosome number.}

\item{PR}{Percentage of population without ROH in a particular window. 100%: no individual with homozygous haplotype in that window.}

\item{population}{Name of the population in quotation marks.}
}
\value{
A table with the RHZ.
}
\description{
This script searches for Regions of Heterozygosity in a population
}
\examples{
# Obtaining RHZ for a group of individuals:
dat<-rhc_data_org(HGDP_cl_Africa)
get_RHZ(dat,1,85,"Africa")
}
