% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{n_vs_sum}
\alias{n_vs_sum}
\title{Number vs Sum of ROH}
\usage{
n_vs_sum(data_1, data_2, color, shape, simul = TRUE)
}
\arguments{
\item{data_1}{The outcome of the function ```roh_sum_id()```.}

\item{data_2}{A data frame with three columns: IID, pop and cont. pop must contain each individual's population, cont must contain each individual's region or continent.}

\item{color}{Factor variable to be introduce as the color guide: ```aes(color=,)```: pop or cont.}

\item{shape}{Factor variable to be introduce as the shape guide: ```aes(shape=,)```: pop or cont.}

\item{simul}{If true simulations of number and sum of ROH for different consanguinity mating is added.}
}
\value{
A figure of the Number vs the Sum of ROH for ROH larger than 1.5Mb.
}
\description{
This script creates a figure of the Number of ROH>=1.5Mb vs. Sum of ROH>=1.5Mb.
It is possible to add the simulated number and sum of ROH for different consanguineous mating.
Simulated data for the number and total length of ROHs (ROH > 1.5 Mb) in the offspring resulting from various consanguineous matings are also displayed.
Points of varying colors represent offspring from different consanguineous relationships: green for second cousins, yellow for first cousins, orange for avuncular (including uncle-niece, aunt-nephew, and double first cousins), and red for incestuous relationships (such as brother-sister and parent-offspring).
Each type of consanguineous mating is depicted through 5,000 simulations.
The dashed diagonal line represents the regression line of N vs S of ROH for two admixed populations from the 1K genomes: ACB and ASW
}
\examples{
# Obtaining basic ROH variables:
rohsum<-roh_summ_id(KGenomes_hom)
pops<-merge(KGenomes_pops,KGenomes_cont,by="pop")
n_vs_sum(rohsum,pops,pop,cont,simul=TRUE)

}
