% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{rhc_data_org}
\alias{rhc_data_org}
\title{RUNS OF HETEROZYGOSITY i}
\usage{
rhc_data_org(POP)
}
\arguments{
\item{POP}{A .hom file from PLINK with all the individuals belonging to the same group or population.}
}
\value{
A file ready to be used to fing RHZ
}
\description{
This script prepares the dataset to search for runs of heterozygosity in the population.
}
