% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{roh_summ_factor}
\alias{roh_summ_factor}
\title{Summarize by factor}
\usage{
roh_summ_factor(data_1, data_2, group_vars)
}
\arguments{
\item{data_1}{The outcome of the function ```roh_summ_id()```.}

\item{data_2}{A file with each individual's factor}

\item{group_vars}{A vector with 1 or more factors do group with.}
}
\value{
A data frame with different variables summarize for each continent.
}
\description{
This script summarize all the outcome of ```roh_summ_id()``` by any factor.
}
\examples{
# Summarizing ROH variables by a general factor:
rohsum<-roh_summ_id(HGDP_hom)
roh_summ_factor(rohsum,HGDP_pops,pop)

}
