% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCT.R
\name{roh_summ_id}
\alias{roh_summ_id}
\title{ROH variables}
\usage{
roh_summ_id(data)
}
\arguments{
\item{data}{A .hom file from PLINK.}
}
\value{
A data frame with different variables.
}
\description{
This script obtains different variables using a PLINK's home file: Sum of ROH>=1.5Mb, Number of ROH>=1.5Mb, Sum of ROH<1.5Mb,
Number of ROH<1.5Mb, Sum of ROH of different sizes, FROH, FoutofROH.
}
